+++ WorkflowFunctionAdminBundle/Resources/config/form.yml
@@ -1,8 +1,8 @@
 parameters:
     open_orchestra_workflow_function_admin.type.workflow_function.class: OpenOrchestra\WorkflowFunctionAdminBundle\Form\Type\WorkflowFunctionType
     open_orchestra_workflow_function_admin.type.workflow_right.class: OpenOrchestra\WorkflowFunctionAdminBundle\Form\Type\WorkflowRightType
-    open_orchestra_workflow_function_admin.type.orchestra_workflow_function.class: OpenOrchestra\WorkflowFunctionAdminBundle\Form\Type\OrchestraWorkflowFunctionType
-    open_orchestra_workflow_function_admin.type.authorization.class: OpenOrchestra\WorkflowFunctionAdminBundle\Form\Type\AuthorizationType
+    open_orchestra_workflow_function_admin.type.orchestra_workflow_function.class: OpenOrchestra\WorkflowFunctionAdminBundle\Form\Type\Component\WorkflowFunctionChoiceType
+    open_orchestra_workflow_function_admin.type.authorization.class: OpenOrchestra\WorkflowFunctionAdminBundle\Form\Type\Component\AuthorizationType
 
 services:
     open_orchestra_workflow_function_admin.type.workflow_function:
@@ -11,24 +11,24 @@ services:
             - %open_orchestra_workflow_function.document.workflow_function.class%
             - @open_orchestra_backoffice.listener.translate_value_initializer
         tags:
-            - { name: form.type, alias: workflow_function}
+            - { name: form.type, alias: oo_workflow_function}
     open_orchestra_workflow_function_admin.type.orchestra_workflow_function:
         class: %open_orchestra_workflow_function_admin.type.orchestra_workflow_function.class%
         arguments:
             - %open_orchestra_workflow_function.document.workflow_function.class%
             - @open_orchestra_backoffice.manager.translation_choice
         tags:
-            - { name: form.type, alias: orchestra_workflow_function }
+            - { name: form.type, alias: oo_workflow_function_choice }
     open_orchestra_workflow_function_admin.type.authorization:
         class: %open_orchestra_workflow_function_admin.type.authorization.class%
         arguments:
             - @open_orchestra_model.repository.content_type
             - %open_orchestra_workflow_function.document.authorization.class%
         tags:
-            - { name: form.type, alias: authorization}
+            - { name: form.type, alias: oo_authorization}
     open_orchestra_workflow_function_admin.type.workflow_right:
         class: %open_orchestra_workflow_function_admin.type.workflow_right.class%
         arguments:
             - %open_orchestra_workflow_function.document.workflow_right.class%
         tags:
-            - { name: form.type, alias: workflow_right}
+            - { name: form.type, alias: oo_workflow_right}
